api_gateway/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
pydantic==1.8.2
sqlalchemy==1.4.22
celery>=5.3.0
numpy==1.26.4
opencv-python==4.7.0.72
pika==1.3.1
python-dotenv==1.0.0
requests==2.25.1
tqdm
urllib3
watchdog==3.0.0
superb-ai-label
pyyaml
# git+https://${GITHUB_TOKEN}@github.com/actnova-inc/actnova-python-sdk.git@main#egg=actnova
pandas
google-api-python-client
redis